add factories

+++ database/migrations/2024_07_23_213217_create_accounts_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
             $table->string('account_number')->unique();
-            $table->decimal('balance', 15, 2)->default(0);
+            $table->bigInteger('balance')->default(0);
             $table->timestamps();
             $table->softDeletes();
         });
